+++ src/test/java/de/stklcode/jvault/connector/HTTPVaultConnectorIT.java
@@ -25,6 +25,7 @@ import de.stklcode.jvault.connector.model.response.*;
 import de.stklcode.jvault.connector.test.Credentials;
 import de.stklcode.jvault.connector.test.VaultConfiguration;
 import org.junit.jupiter.api.*;
+import org.junit.jupiter.api.condition.EnabledIf;
 import org.junit.jupiter.api.io.TempDir;
 
 import java.io.*;
@@ -50,7 +51,7 @@ import static org.junit.jupiter.api.Assumptions.assumeTrue;
  * @since 0.1
  */
 class HTTPVaultConnectorIT {
-    private static String VAULT_VERSION = "1.11.2";  // The vault version this test is supposed to run against.
+    private static String VAULT_VERSION = "1.12.0";  // The vault version this test is supposed to run against.
     private static final String KEY1 = "E38bkCm0VhUvpdCKGQpcohhD9XmcHJ/2hreOSY019Lho";
     private static final String KEY2 = "O5OHwDleY3IiPdgw61cgHlhsrEm6tVJkrxhF6QAnILd1";
     private static final String KEY3 = "mw7Bm3nbt/UWa/juDjjL2EPQ04kiJ0saC5JEXwJvXYsB";
@@ -58,6 +59,7 @@ class HTTPVaultConnectorIT {
     private static final String USER_VALID = "validUser";
     private static final String PASS_VALID = "validPass";
 
+    private static boolean legacy;
     private Process vaultProcess;
     private VaultConnector connector;
 
@@ -67,6 +69,9 @@ class HTTPVaultConnectorIT {
         if (System.getenv("VAULT_VERSION") != null) {
             VAULT_VERSION = System.getenv("VAULT_VERSION");
             System.out.println("Vault version set to " + VAULT_VERSION);
+        }
+        if (compareVersions(VAULT_VERSION, "1.12.0") < 0) {
+            legacy = true;
         }
     }
 
@@ -547,6 +552,8 @@ class HTTPVaultConnectorIT {
 
     @Nested
     @DisplayName("App-ID Tests")
+    @EnabledIf(value = "de.stklcode.jvault.connector.HTTPVaultConnectorIT#isLegacy")
+    @SuppressWarnings("deprecation")
     class AppIdTests {
         private static final String APP_ID = "152AEA38-85FB-47A8-9CBD-612D645BFACA";
         private static final String USER_ID = "5ADF8218-D7FB-4089-9E38-287465DBF37E";
@@ -557,7 +564,6 @@ class HTTPVaultConnectorIT {
         @Test
         @Order(10)
         @DisplayName("Authenticate with App-ID")
-        @SuppressWarnings("deprecation")
         void authAppIdTest() {
             // Try unauthorized access first.
             assumeFalse(connector.isAuthorized());
@@ -580,7 +586,6 @@ class HTTPVaultConnectorIT {
         @Test
         @Order(20)
         @DisplayName("Register App-ID")
-        @SuppressWarnings("deprecation")
         void registerAppIdTest() {
             // Authorize.
             authRoot();
@@ -1074,8 +1079,13 @@ class HTTPVaultConnectorIT {
                     () -> connector.getAuthBackends(),
                     "Could not list supported auth backends"
             );
-            assertEquals(4, supportedBackends.size());
-            assertTrue(supportedBackends.containsAll(List.of(AuthBackend.TOKEN, AuthBackend.USERPASS, AuthBackend.APPID, AuthBackend.APPROLE)));
+            if (legacy) {
+                assertEquals(4, supportedBackends.size());
+                assertTrue(supportedBackends.containsAll(List.of(AuthBackend.TOKEN, AuthBackend.USERPASS, AuthBackend.APPID, AuthBackend.APPROLE)));
+            } else {
+                assertEquals(3, supportedBackends.size());
+                assertTrue(supportedBackends.containsAll(List.of(AuthBackend.TOKEN, AuthBackend.USERPASS, AuthBackend.APPROLE)));
+            }
         }
 
         /**
@@ -1201,7 +1211,11 @@ class HTTPVaultConnectorIT {
      */
     private VaultConfiguration initializeVault(File dir, boolean tls) throws IllegalStateException, IOException {
         File dataDir = new File(dir, "data");
-        copyDirectory(new File(getClass().getResource("/data_dir").getPath()), dataDir);
+        if (legacy) {
+            copyDirectory(new File(getClass().getResource("/data_dir_legacy").getPath()), dataDir);
+        } else {
+            copyDirectory(new File(getClass().getResource("/data_dir").getPath()), dataDir);
+        }
 
         // Generate vault local unencrypted configuration.
         VaultConfiguration config = new VaultConfiguration()
@@ -1294,5 +1308,36 @@ class HTTPVaultConnectorIT {
         StringWriter sw = new StringWriter();
         th.printStackTrace(new PrintWriter(sw, true));
         return sw.getBuffer().toString();
+    }
+
+    /**
+     * Compare two version strings.
+     *
+     * @param version1 Version 1
+     * @param version2 Version 2
+     * @return negative value if version 1 is smaller than version2, positive value of version 1 is greater, 0 if equal
+     */
+    private static int compareVersions(String version1, String version2) {
+        int comparisonResult = 0;
+
+        String[] version1Splits = version1.split("\\.");
+        String[] version2Splits = version2.split("\\.");
+        int maxLengthOfVersionSplits = Math.max(version1Splits.length, version2Splits.length);
+
+        for (int i = 0; i < maxLengthOfVersionSplits; i++) {
+            Integer v1 = i < version1Splits.length ? Integer.parseInt(version1Splits[i]) : 0;
+            Integer v2 = i < version2Splits.length ? Integer.parseInt(version2Splits[i]) : 0;
+            int compare = v1.compareTo(v2);
+            if (compare != 0) {
+                comparisonResult = compare;
+                break;
+            }
+        }
+
+        return comparisonResult;
+    }
+
+    private static boolean isLegacy() {
+        return legacy;
     }
 }